Extract filterByFamiliarity
